+++ src/main/java/spacewars/model/game/elements/bullets/BulletPlayer.java
@@ -2,7 +2,7 @@ package spacewars.model.game.elements.bullets;
 
 public class BulletPlayer extends Bullet
 {
-    public BulletPlayer(int x, int y)
+    public BulletPlayer(double x, double y)
     {
         super(x, y, 30);
     }
