+++ src/main/java/spacewars/view/images/Image.java
@@ -24,7 +24,7 @@ public class Image {
         return ImageIO.read(resourceStream);
     }
 
-    public final void draw(GUI gui, int a, int b) throws IOException {
+    public final void draw(GUI gui, double a, double b) throws IOException {
         for (int y = 0; y < image.getHeight(); y++) {
             for (int x = 0; x < image.getWidth(); x++) {
                 int rgb = image.getRGB(x, y);
